fix(monobank): widen invoice amount fields from int to long
MonobankInvoiceRequest.Amount was int, capping invoices at ~21.5M UAH in kopecks. Widens it plus the response-side MonobankInvoiceStatus.Amount/FinalAmount and MonobankCancelListItem.Amount to long, matching the webhook DTO which already used long. Drops the now-lossy (int) casts at the two invoice-creation sites and in the stub client. Order/Refund DB columns stay int (per-order prices cannot approach the limit).

+++ backend/src/FuelFlow.API/Features/Orders/CreateCheckout/BulkCheckoutCommandHandler.cs
@@ -96,7 +96,7 @@ public sealed class BulkCheckoutCommandHandler
                 new MonobankInvoiceRequest
                 {
                     // Monobank is the one place amounts must be kopecks.
-                    Amount = (int)Money.ToKopecks(totalPrice),
+                    Amount = Money.ToKopecks(totalPrice),
                     MerchantPaymentInfo = $"FuelFlow Bundle",
                     RedirectUrl = _monobankOptions.RedirectUrl,
                     WebhookUrl = _monobankOptions.WebhookUrl
